+++ Models/PurchaseOrderDetails.cs
@@ -5,9 +5,9 @@ namespace PO_Assignment_Project.Models
 {
     public class PurchaseOrderDetails
     {
-        public long ID { get; set; }
+        public int ID { get; set; }
 
-        public long PurchaseOrderID { get; set; }
+        public int PurchaseOrderID { get; set; }
         public PurchaseOrder PurchaseOrder { get; set; }
 
         public int MaterialID { get; set; }
